Social Fabric v2: rootedness B25038 blend, weighted civic, participation mix
- Rootedness: tract 0.6×B25038 long-tenure housing % + 0.4×B07003 same-house 1yr; place same blend; 70/30 tract/place
- Civic: weighted effective node count (1.5/1.0/0.5); Places co-primary for urban_core/suburban
- Participation: 40% refined BMF (N/P/S/W) with legacy fallback, 40% state CPS volunteering proxy, 20% precinct or tract turnout; pillar raw/3.6
- IRS BMF: dual tract counts + stats; runtime counts_mode; build script writes refined+legacy
- Data: volunteering state rates, empty precinct turnout map; DQ cap when participation median fallback

Made-with: Cursor

+++ data_sources/social_fabric_bands.py
@@ -196,7 +196,7 @@ def civic_band_area_type_for_radius(radius_m: int) -> str:
 
 
 def score_civic_gathering_from_bands(
-    count: int,
+    count: float,
     area_type: Optional[str],
     bands: Mapping[str, Any],
     proximity: bool,
